Added overapproximative tests, fixed the one bug they revealed
Added tests that forced some of the bits to top randomly, and then ensured that the resulting abstract value was a correct approximation of the concrete value. This revealed one bug in my "improved" version of the shift operators, having to do with cases where the shift amount was larger than the bit size (e.g. shifting an 8-bit quantity by 9). That bug is now fixed. (That's what I get for "improving" things off the top of my head without testing them...) The code is nearly production-quality at this point, although I still need to make a nice interface for clients to use (or several of them).