Improve correlation performance by nearly 50%
The correlation component was originally architected with incremental support in mind. PowerAnalysis::generateGuesses() allocated an array to store the predictions for each bit in an intermediary cryptographic quantity per subkey. ::analyzeTrace would call that method, and for each trace point, iterate through the guesses in search of a higher correlation. I knew there was overhead involved, but I didn't know how much. Inspired by the performance improvement from optimizing the Hamming weight computation earlier today, I decided to try eliminating the guess precomputations. I wasn't expecting times to go down by nearly 50%! Combined with the Hamming weight optimization, the correlation time is about 1/3 of what it was yesterday. The morals of the story: A) I should profile this component; B) incrementality introduces a significant abstraction penalty, and so should be implemented as a separate component with a compatible interface.